# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface CreativeWorkspace {
  signedIn: boolean;
  /** @minimum 0 */
  unclaimedProjectCount: number;
}

export interface CreativeWorkspaceClaimInput {
  confirm: true;
}

export interface CreativeWorkspaceClaim {
  /** @minimum 0 */
  claimedProjectCount: number;
}

export interface HealthStatus {
  status: string;
}

export interface CreativeBriefInput {
  /**
     * @minLength 20
     * @maxLength 5000
     */
  brief: string;
}

export type CreativeTreatmentGeneratedBy = typeof CreativeTreatmentGeneratedBy[keyof typeof CreativeTreatmentGeneratedBy];


export const CreativeTreatmentGeneratedBy = {
  'google-adk-gemini': 'google-adk-gemini',
} as const;

export interface CreativeTreatment {
  title: string;
  logline: string;
  centralIdea: string;
  emotionalDirection: string;
  tone: string[];
  narrativeApproach: string;
  visualPrinciples: string[];
  audiencePromise: string;
  guardrails: string[];
  generatedBy: CreativeTreatmentGeneratedBy;
}

export type CreativeProjectStatus = typeof CreativeProjectStatus[keyof typeof CreativeProjectStatus];


export const CreativeProjectStatus = {
  generating: 'generating',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface CreativeProject {
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  id: string;
  brief: string;
  status: CreativeProjectStatus;
  treatment: CreativeTreatment | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreativeTreatmentRevision {
  /** @pattern ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$ */
  id: string;
  treatment: CreativeTreatment;
  createdAt: string;
}

export interface ApiError {
  error: string;
}