# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: creative
    description: Google ADK-powered creative development
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /creative-treatment:
    post:
      operationId: createCreativeTreatment
      tags: [creative]
      summary: Create a structured creative treatment
      description: Runs a real Google ADK agent backed by Gemini and saves a treatment to the signed-in account, or to the signed anonymous browser workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreativeBriefInput"
      responses:
        "201":
          description: Creative treatment generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeProject"
        "400":
          description: Missing or invalid creative brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Google AI generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Generation requires a same-origin JSON request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /creative-projects:
    get:
      operationId: listCreativeProjects
      tags: [creative]
      summary: List creative project history
      description: Returns only the signed-in account's projects, or unclaimed projects in the signed anonymous browser workspace, newest first. Responses never expose ownership or internal provider identifiers.
      responses:
        "200":
          description: Creative project history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreativeProject"
  /creative-workspace:
    get:
      operationId: getCreativeWorkspace
      tags: [creative]
      summary: Get the current workspace and available browser projects
      responses:
        "200":
          description: Current ownership mode and number of unclaimed projects in this browser
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeWorkspace"
  /creative-workspace/claim:
    post:
      operationId: claimCreativeWorkspace
      tags: [creative]
      summary: Save this browser's unclaimed projects to the signed-in account
      description: Requires a verified account session and same-origin JSON request. Identity and workspace are resolved exclusively on the server. Transfers all unclaimed projects from the signed workspace cookie atomically, including treatment history. Safe to repeat.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreativeWorkspaceClaimInput"
      responses:
        "200":
          description: Number of projects transferred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeWorkspaceClaim"
        "400":
          description: Invalid claim request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Sign in before claiming projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Cross-origin requests are not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /creative-projects/{projectId}:
    get:
      operationId: getCreativeProject
      tags: [creative]
      summary: Get a creative project
      description: Returns 404 for missing projects and projects owned by another account or browser workspace.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      responses:
        "200":
          description: Creative project and latest validated treatment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeProject"
        "400":
          description: Invalid project identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /creative-projects/{projectId}/treatments:
    get:
      operationId: listCreativeProjectTreatments
      tags: [creative]
      summary: List a project's validated treatment history
      description: History inherits project ownership. Returns 404 for missing projects and projects owned by another account or browser workspace.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      responses:
        "200":
          description: Validated treatment revisions newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreativeTreatmentRevision"
        "400":
          description: Invalid project identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    CreativeWorkspace:
      type: object
      required: [signedIn, unclaimedProjectCount]
      properties:
        signedIn:
          type: boolean
        unclaimedProjectCount:
          type: integer
          minimum: 0
    CreativeWorkspaceClaimInput:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        confirm:
          type: boolean
          enum: [true]
    CreativeWorkspaceClaim:
      type: object
      required: [claimedProjectCount]
      properties:
        claimedProjectCount:
          type: integer
          minimum: 0
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreativeBriefInput:
      type: object
      properties:
        brief:
          type: string
          minLength: 20
          maxLength: 5000
      required:
        - brief
    CreativeTreatment:
      type: object
      properties:
        title:
          type: string
        logline:
          type: string
        centralIdea:
          type: string
        emotionalDirection:
          type: string
        tone:
          type: array
          items:
            type: string
        narrativeApproach:
          type: string
        visualPrinciples:
          type: array
          items:
            type: string
        audiencePromise:
          type: string
        guardrails:
          type: array
          items:
            type: string
        generatedBy:
          type: string
          enum: [google-adk-gemini]
      required:
        - title
        - logline
        - centralIdea
        - emotionalDirection
        - tone
        - narrativeApproach
        - visualPrinciples
        - audiencePromise
        - guardrails
        - generatedBy
    CreativeProject:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
        brief:
          type: string
        status:
          type: string
          enum: [generating, completed, failed]
        treatment:
          oneOf:
            - $ref: "#/components/schemas/CreativeTreatment"
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - brief
        - status
        - treatment
        - createdAt
        - updatedAt
    CreativeTreatmentRevision:
      type: object
      properties:
        id:
          type: string
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
        treatment:
          $ref: "#/components/schemas/CreativeTreatment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - treatment
        - createdAt
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error